refactor: file and folder naming changed to kebab-case
