+++ app/layout.tsx
@@ -4,7 +4,7 @@ import { AppRouterCacheProvider } from '@mui/material-nextjs/v14-appRouter';
 import { Experimental_CssVarsProvider as CssvarsProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import theme from '@/utils/theme';
-import Header from '@/components/Header/Header';
+import Header from '@/components/header/header';
 import Copyright from '@/components/Copyright';
 import { Container } from '@mui/material';
